+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+- Logout endpoint (see [issue](https://gitlab.ilabt.imec.be/prov4itdata-dapsi/web-app/-/issues/53))
+
 ## [0.2.1] - 2021-02-05
 
 ### Changed